refactor: refresh repository structure (#53)
I edited the build interface. I'm no expert with tox so instead I implemented nox (that I know much better). It works the same way, but as the config is done in python it's easier to read.
starting from this new nox builds, I updated the failing pre-commit hooks. now it runs "black", "ruff", "prettier". It covers everything and lead to few modifications (additional docstring and small refactoring)
refresh the tests using pytest-regression and the sphinx builtin pytest mark. It now build a small documentation with 1 single video in vimeo and youtube. I check all the covered builder + "textinfo" for the unsupported node
I refreshed the documentation structure and added a "contributing" section for newomers that may be new to nox.
As everything was set-up I moved the platform related parameter to their respective directives in order to drop the custom node vistors. The only remaining one is because youtube is using a different one for embed and normal. If anybody wants to add a provider it will be much easier
I reorder the methods in utils to make it clearer
I refactor the visitor names to respect the same convention as in sphinx: visit_<node_name>_node_<builder>
I updated the CI as checkout V2 was deprecated and I integrated the pre-commit in case a contributor forget to install them locally

---------

Co-authored-by: David A. Ham <david.ham@imperial.ac.uk>
# Conflicts:
#	pyproject.toml